requirement.txt
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmake==3.26.3
coloredlogs==15.0.1
# cuda-python==12.1.0
Cython==0.29.34
Farama-Notifications==0.0.4
filelock==3.12.0
flatbuffers==23.5.9
fsspec==2023.5.0
gast==0.4.0
google-auth==2.18.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
gymnasium==0.29.0
h5py==3.8.0
huggingface-hub==0.14.1
humanfriendly==10.0
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.6.0
jax==0.4.10
Jinja2==3.1.2
joblib==1.3.2
keras==2.12.0
libclang==16.0.0
lit==16.0.3
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.1.0
mpmath==1.3.0
networkx==3.1
numpy==1.23.5
# nvidia-cublas-cu11==11.10.3.66
# nvidia-cuda-cupti-cu11==11.7.101
# nvidia-cuda-nvrtc-cu11==11.7.99
# nvidia-cuda-runtime-cu11==11.7.99
# nvidia-cudnn-cu11==8.5.0.96
# nvidia-cufft-cu11==10.9.0.58
# nvidia-curand-cu11==10.2.10.91
# nvidia-cusolver-cu11==11.4.0.1
# nvidia-cusparse-cu11==11.7.4.91
# nvidia-nccl-cu11==2.14.3
# nvidia-nvtx-cu11==11.7.91
# nvtx==0.2.5
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.14.0
onnxruntime==1.14.1
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
protobuf==4.23.0
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
sacremoses==0.0.53
scikit-learn==1.3.0
scipy==1.10.1
sentencepiece==0.1.99
six==1.16.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.14.0
# tensorboard-data-server==0.7.0
# tensorflow==2.12.0
# tensorflow-estimator==2.12.0
# tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torch-geometric==2.3.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
transformers==4.29.1
# triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.3.4
wrapt==1.14.1
zipp==3.15.0
